GameDetail Screen calls working

+++ app/src/main/java/com/sgut/android/nationalfootballleague/data/remote/responses/game_details/AwayTeamOdds.kt
@@ -7,8 +7,8 @@ data class AwayTeamOdds (
 
   @SerializedName("favorite"   ) var favorite   : Boolean? = null,
   @SerializedName("underdog"   ) var underdog   : Boolean? = null,
-  @SerializedName("moneyLine"  ) var moneyLine  : Int?     = null,
-  @SerializedName("spreadOdds" ) var spreadOdds : Int?     = null,
+  @SerializedName("moneyLine"  ) var moneyLine  : Float?     = null,
+  @SerializedName("spreadOdds" ) var spreadOdds : Float?     = null,
   @SerializedName("teamId"     ) var teamId     : String?  = null
 
 )
