+++ app/src/main/java/com/sgut/android/nationalfootballleague/data/remote/responses/game_details/Pickcenter.kt
@@ -7,8 +7,8 @@ data class Pickcenter (
 
   @SerializedName("provider"     ) var provider     : Provider?         = Provider(),
   @SerializedName("details"      ) var details      : String?           = null,
-  @SerializedName("overUnder"    ) var overUnder    : Int?              = null,
-  @SerializedName("spread"       ) var spread       : Int?              = null,
+  @SerializedName("overUnder"    ) var overUnder    : Float?              = null,
+  @SerializedName("spread"       ) var spread       : Float?              = null,
   @SerializedName("awayTeamOdds" ) var awayTeamOdds : AwayTeamOdds?     = AwayTeamOdds(),
   @SerializedName("homeTeamOdds" ) var homeTeamOdds : HomeTeamOdds?     = HomeTeamOdds(),
   @SerializedName("links"        ) var links        : ArrayList<String> = arrayListOf()
